vl: fix low strength denoise filter
A filter strength of zero or one doesn't make any
sense. Thanks to Andy Furniss for pointing this out.

+++ src/gallium/state_trackers/vdpau/vdpau_private.h
@@ -311,7 +311,7 @@ typedef struct
 
    struct {
       bool supported, enabled;
-      float level;
+      unsigned level;
       struct vl_median_filter *filter;
    } noise_reduction;
